Data.BiMap

Documentation

data BiMap v a b where

Constructors

MkBiMap :: v ~ b => !(Map a b) -> !(Map b (Set a)) -> BiMap v a b

biMapToMap :: BiMap v a b -> Map a b

biMapFromMap :: (Ord k, Ord v) => Map k v -> BiMap v k v

decodeMapAsBimap :: (FromCBOR a, FromCBOR b, Ord a, Ord b) => Decoder s (BiMap b a b)

Decode a serialised CBOR Map as a Bimap

addBack :: (Ord v, Ord k) => v -> k -> Map v (Set k) -> Map v (Set k)

retract :: (Ord v, Ord k) => v -> k -> Map v (Set k) -> Map v (Set k)

insertBackwards :: (Ord k, Ord v) => v -> v -> k -> Map v (Set k) -> Map v (Set k)

insertWithBiMap :: (Ord k, Ord v) => (v -> v -> v) -> k -> v -> BiMap v k v -> BiMap v k v

biMapEmpty :: BiMap v k v

biMapFromList :: (Ord k, Ord v) => (v -> v -> v) -> [(k, v)] -> BiMap v k v

mapflip :: (v -> v -> v) -> v -> v -> v

biMapFromAscDistinctList :: (Ord k, Ord v) => [(k, v)] -> BiMap v k v

Warning - invariant that keys are distinct and in ascending order is not checked. Make sure it is not violated, otherwise crazy things will happen.

type Bimap k v = BiMap v k v

removeval :: (Ord k, Ord v) => v -> BiMap v k v -> BiMap v k v