Data.MapExtras

Description

Sometimes we need to write our own version of functions over Map that do not appear in the "containers" library. This module is for such functions.

For example:

Version of withoutKeys where both arguments are Map

Comparing that two maps have exactly the same set of keys

The intersection of two maps guarded by a predicate.

((dom stkcred) ◁ deleg) ▷ (dom stpool)) ==>
intersectDomP (\ k v -> Map.member v stpool) stkcred deleg

Documentation

data StrictTriple a b c

Constructors

StrictTriple !a !b !c

extract :: Ord k => k -> Map k b -> (Maybe b, Map k b)

Just like delete, but also returns the value if it was indeed deleted from the map.

noKeys :: Ord k => Map k a -> Map k b -> Map k a

keysEqual :: Ord k => Map k v1 -> Map k v2 -> Bool

splitMemberMap :: Ord k => k -> Map k a -> StrictTriple (Map k a) Bool (Map k a)

A variant of splitLookup that indicates only whether the key was present, rather than producing its value. This is used to implement keysEqual to avoid allocating unnecessary Just constructors.

Note - this is a copy pasted internal function from "containers" package adjusted to return StrictTriple

splitMemberSet :: Ord a => a -> Set a -> StrictTriple (Set a) Bool (Set a)

O(log n). Performs a split but also returns whether the pivot element was found in the original set.

This is a modified version of splitMember, where StrictTriple is used instead of a lazy one for minor performance gain.

intersectDomP :: Ord k => (k -> v2 -> Bool) -> Map k v1 -> Map k v2 -> Map k v2

intersetDomP p m1 m2 == Keep the key and value from m2, iff (the key is in the dom of m1) && ((p key value) is true)

intersectDomPLeft :: Ord k => (k -> v2 -> Bool) -> Map k v1 -> Map k v2 -> Map k v1

Similar to intersectDomP, except the Map returned has the same key as the first input map, rather than the second input map.

intersectMapSetFold :: Ord k => (k -> v -> ans -> ans) -> Map k v -> Set k -> ans -> ans

fold over the intersection of a Map and a Set

disjointMapSetFold :: Ord k => (k -> v -> ans -> ans) -> Map k v -> Set k -> ans -> ans

Fold with accum all those pairs in the map, not appearing in the set.

extractKeys :: Ord k => Map k a -> Set k -> (Map k a, Map k a)

Partition the Map according to keys in the Set. This is equivalent to:

extractKeys m s === (withoutKeys m s, restrictKeys m s)

extractKeysSmallSet :: Ord k => Map k a -> Set k -> (Map k a, Map k a)

It has been discovered expirementally through benchmarks that for small Set size of under around 6 elements this function performs faster than extractKeys#

fromKeys :: (Foldable f, Ord k) => (k -> v) -> f k -> Map k v

Convert any foldable data structure with keys to a Map. Implemented in terms of fromList, therefore last duplicate key wins.

fromElems

Arguments

:: (Foldable f, Ord k)

=> (v -> k)
Function that will create a key from a value. Most common case is a hashing function.

-> f v

-> Map k v

Convert any foldable data structure with values to a Map. Implemented in terms of fromList, therefore last duplicate key wins.