Data.Roundtrip

Description

Defines reusable abstractions for testing Roundtrip properties of CBOR instances

Documentation

roundTrip :: (ToCBOR t, FromCBOR t) => t -> RoundTripResult t

roundTrip' :: (t -> Encoding) -> (forall s. Decoder s t) -> t -> RoundTripResult t

embedTrip :: (ToCBOR t, FromCBOR s) => t -> RoundTripResult s

Can we serialise a type, and then deserialise it as something else?

embedTrip' :: (s -> Encoding) -> (forall x. Decoder x t) -> s -> RoundTripResult t

roundTripAnn :: (ToCBOR t, FromCBOR (Annotator t)) => t -> RoundTripResult t

embedTripAnn :: forall s t. (ToCBOR t, FromCBOR (Annotator s)) => t -> RoundTripResult s

type RoundTripResult t = Either DeserialiseFailure (ByteString, t)