Data.UMap

Documentation

data Trip coin ptr pool where

Bundled Patterns

pattern Triple :: StrictMaybe coin -> Set ptr -> StrictMaybe pool -> Trip coin ptr pool

tripReward :: Trip coin ptr pool -> Maybe coin

tripRewardActiveDelegation :: Trip coin ptr pool -> Maybe coin

tripDelegation :: Trip coin ptr pool -> Maybe pool

data UMap coin cred pool ptr

Constructors

UnifiedMap !(Map cred (Trip coin ptr pool)) !(Map ptr cred)

class UnifiedView coin cred pool ptr k v where

Methods

tag :: Tag coin cred pool ptr k v

umInvariant :: (Ord cred, Ord ptr) => cred -> ptr -> UMap coin cred pool ptr -> Bool

It is worthwhie stating the invariant that holds on a Unified Map The ptrmap and the ptrT field of the tripmap are inverses.

unView :: View coin cr pl ptr k v -> UMap coin cr pl ptr

unUnify :: View coin cred pool ptr k v -> Map k v

This is expensive, use it wisely (like maybe once per epoch boundary to make a SnapShot) See also domRestrictedView, which domain restricts before computing a view.

viewToVMap :: Ord cred => View coin cred pool ptr k v -> VMap VB VB k v

This is expensive, use it wisely (like maybe once per epoch boundary to make a SnapShot)

rewView :: UMap coin cred pool ptr -> Map cred coin

delView :: UMap coin cred pool ptr -> Map cred pool

ptrView :: UMap coin cred pool ptr -> Map ptr cred

domRestrictedView :: (Ord ptr, Ord cred) => Set k -> View coin cred pl ptr k v -> Map k v

Return the appropriate View of a domain restricted Umap. f setk is small this should be efficient.

zero :: Trip coin ptr pool -> Bool

Is there no information in a Triple? If so then we can delete it from the UnifedMap

zeroMaybe :: Trip coin ptr pool -> Maybe (Trip coin ptr pool)

mapNext :: Map k v -> Maybe (k, v, Map k v)

mapLub :: Ord k => k -> Map k v -> Maybe (k, v, Map k v)

next :: View coin cr pl ptr k v -> Maybe (k, v, View coin cr pl ptr k v)

leastUpperBound :: (Ord ptr, Ord cr) => k -> View coin cr pool ptr k v -> Maybe (k, v, View coin cr pool ptr k v)

empty :: UMap coin cr pool ptr

delete :: (Ord cr, Ord ptr) => k -> View coin cr pool ptr k v -> UMap coin cr pool ptr

delete' :: (Ord cr, Ord ptr) => k -> View coin cr pool ptr k v -> View coin cr pool ptr k v

insertWith :: (Ord cr, Monoid coin, Ord ptr) => (v -> v -> v) -> k -> v -> View coin cr pool ptr k v -> UMap coin cr pool ptr

insertWith' :: (Ord cr, Monoid coin, Ord ptr) => (v -> v -> v) -> k -> v -> View coin cr pool ptr k v -> View coin cr pool ptr k v

Special insertion:

Keeps the value already in the ViewMap if the key k is already there:

insertWith' (\ old new -> old) k v xs

Replaces the value already in the ViewMap with v, if key k is already there:

insertWith' (\ old new -> new) k v xs

Replaces the value already in the ViewMap with the sum, if key k is already there:

insertWith' (\ old new -> old+new) k v xs

Ignores combine if the key k is NOT already in the ViewMap, and inserts v:

insertWith' combine k v xs

insert :: (Ord cr, Monoid coin, Ord ptr) => k -> v -> View coin cr pool ptr k v -> UMap coin cr pool ptr

insert' :: (Ord cr, Monoid coin, Ord ptr) => k -> v -> View coin cr pool ptr k v -> View coin cr pool ptr k v

lookup :: (Ord cr, Ord ptr) => k -> View coin cr pool ptr k v -> Maybe v

isNull :: View coin cr pool ptr k v -> Bool

domain :: Ord cr => View coin cr pool ptr k v -> Set k

range :: (Ord coin, Ord pool, Ord cr) => View coin cr pool ptr k v -> Set v

(∪) :: (Ord cr, Monoid coin, Ord ptr) => View coin cr pool ptr k v -> (k, v) -> UMap coin cr pool ptr

Union with left preference, so if k, already exists, do nothing, if it doesn't exist insert it.

(⨃) :: (Ord cr, Monoid coin, Ord ptr) => View coin cr pool ptr k v -> Map k v -> UMap coin cr pool ptr

Union with right preference, so if k, already exists, then its value is overwritten with v

(∪+) :: (Ord cred, Monoid coin) => View coin cred pool ptr k coin -> Map k coin -> UMap coin cred pool ptr

(⋪) :: (Ord cred, Ord ptr) => Set k -> View coin cred pool ptr k v -> UMap coin cred pool ptr

(⋫) :: (Ord cred, Ord ptr, Ord coin, Ord pool) => View coin cred pool ptr k v -> Set v -> UMap coin cred pool ptr

This is slow for Delegations and Rewards Views, better hope they are small

member :: (Ord cr, Ord ptr) => k -> View coin cr pool ptr k v -> Bool

notMember :: (Ord cr, Ord ptr) => k -> View coin cr pool ptr k v -> Bool

domRestrict :: (Ord cr, Ord ptr) => View coin cr pool ptr k v -> Map k u -> Map k u

data Tag coin cred pool ptr k v where

Constructors

Rew :: Tag coin cred pool ptr cred coin

Del :: Tag coin cred pool ptr cred pool

Ptr :: Tag coin cred pool ptr ptr cred

data View coin cr pl ptr k v where

Constructors

Rewards :: !(UMap coin cr pl ptr) -> View coin cr pl ptr cr coin

Delegations :: !(UMap coin cr pl ptr) -> View coin cr pl ptr cr pl

Ptrs :: !(UMap coin cr pl ptr) -> View coin cr pl ptr ptr cr

findWithDefault :: (Ord cred, Ord ptr) => a -> k -> View coin cred pool ptr k a -> a

size :: View coin cred pool ptr k a -> Int

A View is a view, so the size of the view is NOT the same as the size of the underlying triple map.

unify :: (Monoid coin, Ord cred, Ord ptr) => Map cred coin -> Map cred pool -> Map ptr cred -> UMap coin cred pool ptr

Create a UMap from 3 separate maps. For use in tests only.