Data.CanonicalMaps

Documentation

class Eq t => CanonicalZero t where

Methods

zeroC :: t

joinC :: t -> t -> t

canonicalInsert :: (Ord k, CanonicalZero a) => (a -> a -> a) -> k -> a -> Map k a -> Map k a

canonicalMapUnion :: (Ord k, CanonicalZero a) => (a -> a -> a) -> Map k a -> Map k a -> Map k a

canonicalMap :: (Ord k, CanonicalZero a) => (a -> a) -> Map k a -> Map k a

pointWise :: (Ord k, CanonicalZero v) => (v -> v -> Bool) -> Map k v -> Map k v -> Bool

data Map k a

A Map from keys k to values a.

The Semigroup operation for Map is union, which prefers values from the left operand. If m1 maps a key k to a value a1, and m2 maps the same key to a different value a2, then their union m1 <> m2 maps k to a1.