Data.OMap.Strict

Documentation

class Ord k ⇒ HasOKey k v | v → k where

Class of types that can be mapped by a lens or a projection to an Ord type.

For a type V, defines a lens from V to and Ord type K.

Methods

okeyL ∷ Lens' v k

data OMap k v where

A general-purpose finite, insert-ordered, map that is strict in its keys and values.

The strictness is enforced by the underlying strict Map that can be looked-up by a projection or lens. and the ordering is maintained by the constructing functions, leveraging StrictSeq to hold the insert-order of the keys.

TODO: DecShareCBOR instance

Bundled Patterns

pattern Empty ∷ OMap k v
\(O(1)\)

pattern (:<|:) ∷ (HasOKey k v, Ord k) ⇒ v → OMap k v → OMap k v infixr 5
\(O(\log n)\).

pattern (:|>:) ∷ (HasOKey k v, Ord k) ⇒ OMap k v → v → OMap k v infixl 5
\(O(\log n)\).

null ∷ OMap k v → Bool

\(O(1)\).

size ∷ OMap k v → Int

\(O(1)\).

empty ∷ OMap k v

\(O(1)\).

singleton ∷ HasOKey k v ⇒ v → OMap k v

\(O(1)\). Strict in its arguments.

lookup ∷ Ord k ⇒ k → OMap k v → Maybe v

\(O(\log n)\). If the key is not present lookup returns Nothing.

member ∷ Ord k ⇒ k → OMap k v → Bool

\(O(\log n)\). Key membership check.

(!?) ∷ Ord k ⇒ OMap k v → k → Maybe v

fliped version of lookup

mapUnsafe ∷ (v1 → v2) → OMap k v1 → OMap k v2

This mapping function is only safe when the key stored in the new value matches the key stored in the new value. This invariant is not checked for performance reasons

fromSet ∷ HasOKey k v ⇒ Set v → OMap k v

\(O(n \log n)\).

fromFoldable ∷ (Foldable f, HasOKey k v) ⇒ f v → OMap k v

\(O(n \log n)\). Checks membership before snoc'ing. De-duplicates the StrictSeq without overwriting. Starts from the left or head, using foldl'

fromFoldableDuplicates ∷ (Foldable f, HasOKey k v, Ord v) ⇒ f v → (Set v, OMap k v)

\(O(n \log n)\). Checks membership before snoc'ing. De-duplicates the StrictSeq and collects and returns the duplicates found. Starts from the left or head, using foldl'

toMap ∷ OMap k v → Map k v

\(O(1)\).

assocList ∷ Ord k ⇒ OMap k v → [(k, v)]

elems ∷ Ord k ⇒ OMap k v → [v]

toStrictSeq ∷ Ord k ⇒ OMap k v → StrictSeq v

\(O(n \log n)\).

toStrictSeqOKeys ∷ OMap k v → StrictSeq k

\(O(1)\).

toStrictSeqOfPairs ∷ Ord k ⇒ OMap k v → StrictSeq (k, v)

\(O(n \log n)\).

invariantHolds ∷ OMap k v → Bool

\(O(1)\). Shallow invariant using just length and size.

invariantHolds' ∷ Ord k ⇒ OMap k v → Bool

\(O(n \log n)\). Deep, costly invariant using membership check for each value. By the pigeon-hole principle, this check is exhaustive.

(|>) ∷ HasOKey k v ⇒ OMap k v → v → OMap k v infixl 5

\(O(\log n)\). Checks membership before snoc'ing.

(<|) ∷ HasOKey k v ⇒ v → OMap k v → OMap k v infixr 5

\(O(\log n)\). Checks membership before cons'ing.

(<||) ∷ HasOKey k v ⇒ v → OMap k v → OMap k v infixr 5

\(O(\log n)\). Checks membership before cons'ing. Overwrites a duplicate.

(||>) ∷ HasOKey k v ⇒ OMap k v → v → OMap k v infixl 5

\(O(\log n)\). Checks membership before snoc'ing. Overwrites a duplicate.

(|><) ∷ HasOKey k v ⇒ OMap k v → OMap k v → OMap k v infixl 5

\( O(n \log m) \). For every uncons-ed element from the sequence on the right, check its membership in the sequence on the left, before snoc'ing it. Preserve order. Remove duplicates from sequence on the right.

(><|) ∷ HasOKey k v ⇒ OMap k v → OMap k v → OMap k v infixr 5

\( O(m \log n) \). For every unsnoc-ed element from the sequence on the left, check its membership in the sequence on the right, before cons'ing it. Preserve order. Remove duplicates from sequence on the left.

elem ∷ (HasOKey k v, Eq v) ⇒ v → OMap k v → Bool

\(O(\log n)\). Value membership check.

extractKeys ∷ Ord k ⇒ Set k → OMap k v → (OMap k v, Map k v)

\(O(n)\). Given a Set of ks, and an OMap k v return a pair of Map and OMap where the ks in the Set have been removed from the OMap and presented as a separate Map.

adjust ∷ HasOKey k v ⇒ (v → v) → k → OMap k v → OMap k v

\(O(n)\). Like adjust.

Returns the original OMap unaltered when the key does not exist.

If the key exists, then the function is applied to the value, but we need to consider three possible cases:

The modified value's okeyL is unaltered - we return omap with the adjusted value,

The modified value's okeyL is altered, but not a duplicate - we return the omap with adjusted key (in place) and value

The modified value's okeyL is altered and is a duplicate - we return the omap with the old key deleted from the sequence but without inserting the new key since it is a duplicate, and deleting old value and inserting the new value in place of its duplicate.

Examples:

>>> :set -XFlexibleInstances -XMultiParamTypeClasses
>>> import Data.OMap.Strict
>>> import Lens.Micro
>>> instance HasOKey Int (Int, Char) where okeyL = _1
>>> let m = fromFoldable $ zip [1,2] ['a','b'] :: OMap Int (Int, Char)
>>> m
StrictSeq {fromStrict = fromList [(1,(1,'a')),(2,(2,'b'))]}
>>> let adjustingFn (k, v) = (k, succ v) -- Changes the value
>>> let overwritingAdjustingFn (k,v) = (succ k, v) -- Changes the `okeyL`.
>>> adjust adjustingFn 1 m
StrictSeq {fromStrict = fromList [(1,(1,'b')),(2,(2,'b'))]}
>>> adjust overwritingAdjustingFn  1 m
StrictSeq {fromStrict = fromList [(2,(2,'a'))]}

filter ∷ Ord k ⇒ (v → Bool) → OMap k v → OMap k v

\( O(n \log n) \)

decodeOMap ∷ HasOKey k v ⇒ Decoder s v → Decoder s (OMap k v)