Data.OSet.Strict

Documentation

data OSet a where

A general-purpose finite, ordered, set that is strict in its values.

The strictness is enforced by the underlying StrictSeq from base, and the uniqueness of the values is enforced in this module, by the constructing functions by leveraging an accompanying Set.

TODO: @aniketd Implement DecShareCBOR

Bundled Patterns

pattern Empty :: OSet a

pattern (:<|:) :: StrictSeq a)) :*: S1 ('MetaSel ('Just "osSet") 'NoSourceUnpackedness 'SourceStrict 'DecidedStrict) (Rec0 (Set a))))

type Item (OSet a)

Instance details

Defined in Data.OSet.Strict

type Item (OSet a) = a

null :: OSet a -> Bool

\( O(1) \).

size :: OSet a -> Int

\( O(1) \).

empty :: OSet a

\( O(1) \).

singleton :: a -> OSet a

\( O(1) \). Strict in its argument.

lookup :: Int -> OSet a -> Maybe a

\( O(\log(\min(i,n-i))) \). The element at the specified position, counting from 0. If the specified position is negative or at least the length of the sequence, lookup returns Nothing.

member :: Ord a => a -> OSet a -> Bool

\(O(\log n)\). Membership check.

(!?) :: OSet a -> Int -> Maybe a

fliped version of lookup

fromList :: Ord a => [a] -> OSet a

\(O(n \log n)\). Convert to an OSet from a List.

fromStrictSeq :: Ord a => StrictSeq a -> OSet a

\(O(n \log n)\). Checks membership before snoc-ing. De-duplicates the StrictSeq without overwriting.

fromStrictSeqDuplicates :: Ord a => StrictSeq a -> (Set a, OSet a)

\(O(n \log n)\). Checks membership before snoc-ing. Returns a 2-tuple, with fst as a Set of duplicates found and the snd as the de-duplicated OSet without overwriting. Starts from the left or head, using foldl'

toStrictSeq :: OSet a -> StrictSeq a

\( O(1) \) - Extract underlying strict sequence

toSet :: OSet a -> Set a

\( O(1) \) - Extract underlying Set

fromSet :: Set a -> OSet a

\( O(n) \).

fromFoldable :: (Foldable f, Ord a) => f a -> OSet a

Using a Foldable instance of the source data structure convert it to an OSet

fromFoldableDuplicates :: (Foldable f, Ord a) => f a -> (Set a, OSet a)

\(O(n \log n)\). Checks membership before snoc-ing. Returns a 2-tuple, with fst as a Set of duplicates found and the snd as the de-duplicated OSet without overwriting.

invariantHolds :: OSet a -> Bool

\( O(1) \). Shallow invariant using just length and size.

invariantHolds' :: Ord a => OSet a -> Bool

\( O(n \log n) \). Deep invariant using set membership check for each value. By the pigeon-hole principle, this check is exhaustive.

(|>) :: Ord a => OSet a -> a -> OSet a infixl 5

(<|) :: Ord a => a -> OSet a -> OSet a infixr 5

(|><) :: Ord a => OSet a -> OSet a -> OSet a infixl 5

\( O(n\log(m*n)) \). For every uncons-ed element from the sequence on the right, checks its membership in the sequence on the left, before snoc'ing it. Preserves order. Remove duplicates from sequence on the right.

(><|) :: Ord a => OSet a -> OSet a -> OSet a infixr 5

\( O(m\log(m+n)) \). For every unsnoc-ed element from the sequence on the left, checks its membership in the sequence on the right, before cons'ing it. Preserves order. Remove duplicates from sequence on the left.

filter :: Ord a => (a -> Bool) -> OSet a -> OSet a

\( O(n \log n) \)